+++ jgulp/tasks/release/redshop.js
@@ -35,7 +35,6 @@ function getIncludedExtensions() {
         "./plugins/redshop_export/shopper_group_product_price/**",
         "./plugins/redshop_export/user/**",
         "./plugins/redshop_export/order/**",
-        "./plugins/redshop_export/order_item/**",
         "./plugins/redshop_import/attribute/**",
         "./plugins/redshop_import/category/**",
         "./plugins/redshop_import/field/**",
