--- plugins/redshop_export/order_item/language/en-GB/en-GB.plg_redshop_export_order_item.sys.ini
@@ -1,4 +1,0 @@
-PLG_REDSHOP_EXPORT_ORDER_ITEM="redSHOP Export: Order Item"
-PLG_REDSHOP_EXPORT_ORDER_ITEM_DESC="Plugin for provide export Order Item of redSHOP"
-
-PLG_REDSHOP_EXPORT_ORDER_ITEM_TITLE="Orders Item"
