--- plugins/redshop_export/order_item/order_item.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<extension version="3.0" type="plugin" group="redshop_export" method="upgrade">
-    <name>PLG_REDSHOP_EXPORT_ORDER_ITEM</name>
-    <version>1.0.0</version>
-    <redshop>2.1.2.2</redshop>
-    <creationDate>June 2017</creationDate>
-    <author>redCOMPONENT.com</author>
-    <authorEmail>email@redcomponent.com</authorEmail>
-    <authorUrl>www.redcomponent.com</authorUrl>
-    <copyright>(C) 2017 redCOMPONENT.com</copyright>
-    <license>GNU/GPL</license>
-    <description>PLG_REDSHOP_EXPORT_ORDER_ITEM_DESC</description>
-    <files>
-        <filename plugin="order_item">order_item.php</filename>
-        <filename>index.html</filename>
-        <folder>language</folder>
-    </files>
-</extension>
